refactor: fixes unnecessary casts

+++ test/unit/UsdnProtocol/Actions/Liquidation.t.sol
@@ -470,7 +470,7 @@ contract TestUsdnProtocolLiquidation is UsdnProtocolBaseFixture {
         int256 collateralLiquidated = protocol.i_tickValue(
             posId.tick,
             price,
-            uint256(protocol.getLongTradingExpo(price)),
+            protocol.getLongTradingExpo(price),
             protocol.getLiqMultiplierAccumulator(),
             protocol.getTickData(posId.tick)
         );
@@ -547,7 +547,7 @@ contract TestUsdnProtocolLiquidation is UsdnProtocolBaseFixture {
         int256 collateralLiquidated = protocol.i_tickValue(
             posId.tick,
             price,
-            uint256(protocol.getLongTradingExpo(price)),
+            protocol.getLongTradingExpo(price),
             protocol.getLiqMultiplierAccumulator(),
             protocol.getTickData(posId.tick)
         );
