+++ test/unit/UsdnProtocol/Actions/ValidateOpenPosition.t.sol
@@ -377,7 +377,7 @@ contract TestUsdnProtocolActionsValidateOpenPosition is UsdnProtocolBaseFixture 
         (data.validateTick,) = protocol.i_getTickFromDesiredLiqPrice(
             protocol.i_getLiquidationPrice(data.validatePrice, uint128(protocol.getMaxLeverage())),
             data.validatePrice,
-            uint256(protocol.getLongTradingExpo(data.validatePrice)),
+            protocol.getLongTradingExpo(data.validatePrice),
             protocol.getLiqMultiplierAccumulator(),
             protocol.getTickSpacing(),
             data.originalLiqPenalty
